move collector tests to TCK
