+++ vertx-mysql-client/src/test/java/io/vertx/mysqlclient/tck/MySQLCollectorTest.java
@@ -1,0 +1,18 @@
+package io.vertx.mysqlclient.tck;
+
+import io.vertx.ext.unit.junit.VertxUnitRunner;
+import io.vertx.mysqlclient.junit.MySQLRule;
+import io.vertx.sqlclient.tck.CollectorTestBase;
+import org.junit.ClassRule;
+import org.junit.runner.RunWith;
+
+@RunWith(VertxUnitRunner.class)
+public class MySQLCollectorTest extends CollectorTestBase {
+  @ClassRule
+  public static MySQLRule rule = MySQLRule.SHARED_INSTANCE;
+
+  @Override
+  protected void initConnector() {
+    connector = ClientConfig.CONNECT.connect(vertx, rule.options());
+  }
+}
