+++ vertx-pg-client/src/test/java/io/vertx/pgclient/tck/PgCollectorTest.java
@@ -1,0 +1,18 @@
+package io.vertx.pgclient.tck;
+
+import io.vertx.ext.unit.junit.VertxUnitRunner;
+import io.vertx.pgclient.junit.PgRule;
+import io.vertx.sqlclient.tck.CollectorTestBase;
+import org.junit.ClassRule;
+import org.junit.runner.RunWith;
+
+@RunWith(VertxUnitRunner.class)
+public class PgCollectorTest extends CollectorTestBase {
+  @ClassRule
+  public static PgRule rule = new PgRule();
+
+  @Override
+  protected void initConnector() {
+    connector = ClientConfig.CONNECT.connect(vertx, rule.options());
+  }
+}
